backend/pyproject.toml
[tool.poetry]
name = "linguify"
version = "0.1.0"
description = ""
authors = []
readme = "Readme.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "5.1.10"
djangorestframework = "3.16.0"
django-filter = "25.1"
pillow = "11.2.1"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"
pytest-django = "^4.9.0"
# psycopg2 = "^2.9.10"
openpyxl = "^3.1.5"
channels = "^4.2.0"
psycopg2-binary = "^2.9.10"
requests = "^2.32.4"
pandas = "^2.2.3"

# AI providers - optional dependencies
openai = "^1.88.0"  # Pour l'intégration OpenAI
drf-spectacular = "^0.28.0"
redis = "^6.2.0"
pyjwt = "^2.10.1"
whitenoise = "^6.9.0"
pywebpush = "^2.0.3"

# Voice/Audio dependencies
pyttsx3 = "^2.90"
SpeechRecognition = "^3.10.0"
# pyaudio = "^0.2.13"  # Optional - difficile à installer sur Windows
supabase = "^2.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"